feat: carry uint64 as Avro fixed(8) rather than long, which lets the sync cursors keep their unsigned type instead of being narrowed, name every event field after its json tag so the schema and the JSON wire agree, and populate every field in the round-trip test because zero values are why a uint64 mapped to long passed in the first place

+++ internal/app/worker/wmail/imap_google.go
@@ -14,7 +14,7 @@ func (w *WMail) ImapGoogleSync(ctx context.Context, lastHistoryID uint64) *errx.
 		if err := w.onEvent(models.JobEventTypeHistoryIDUpdate, &models.JobEventHistoryIDUpdate{
 			UserID:    w.UserID,
 			EmailID:   w.ID,
-			HistoryID: int64(newHistoryID),
+			HistoryID: newHistoryID,
 		}); err != nil {
 			return nil
 		}
